chore: separate cairo0 and cairo1 test data
