+++ starknet-core/test-data/contracts/cairo0/artifacts/deployable.hashes.json
