+++ starknet-core/test-data/contracts/cairo0/artifacts/deployable.txt
