+++ starknet-core/test-data/contracts/cairo0/artifacts/event_example.hashes.json
