+++ starknet-core/test-data/contracts/cairo0/artifacts/oz_account.txt
