+++ starknet-core/test-data/contracts/cairo0/contracts/Deployable.cairo
