+++ starknet-core/test-data/contracts/cairo0/contracts/EventExample.cairo
