+++ starknet-core/test-data/contracts/cairo0/docker_entry.sh
