+++ starknet-core/test-data/contracts/cairo0/generate_artifacts.sh
