+++ starknet-core/test-data/contracts/cairo0/scripts/generate_hashes.py
