+++ starknet-core/test-data/contracts/cairo1/artifacts/erc20_sierra.txt
